api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic>=2.7.4,<3.0.0
python-dotenv
biopython
pandas
pydantic
typing
email-validator